Split of input/output functionalities
